#SB-210: Create a service that handles the management of artifacts and extract the related code from the ArtifactRestlet into it
- Extracted the code.
- Minor refactoring.
