+++ strongbox-commons/src/main/java/org/carlspring/strongbox/security/encryption/EncryptionConstants.java
