+++ strongbox-commons/src/main/java/org/carlspring/strongbox/util/ArtifactFileUtils.java
