+++ strongbox-storage/strongbox-storage-api/src/main/java/org/carlspring/strongbox/storage/services/VersionValidatorService.java
@@ -1,0 +1,15 @@
+package org.carlspring.strongbox.storage.services;
+
+import org.carlspring.strongbox.storage.validation.version.VersionValidator;
+
+import java.util.Set;
+
+/**
+ * @author mtodorov
+ */
+public interface VersionValidatorService
+{
+
+    Set<VersionValidator> getValidators();
+
+}
