+++ strongbox-testing/strongbox-testing-storage-resources/strongbox-testing-storage-indexing-resources/src/main/resources/META-INF/spring/strongbox-storage-indexing-context.xml
@@ -9,4 +9,6 @@
 
     <bean id="repositoryIndexManager" class="org.carlspring.strongbox.storage.indexing.RepositoryIndexManager"/>
 
+    <bean id="artifactManagementService" class="org.carlspring.strongbox.storage.services.impl.ArtifactManagementServiceImpl"/>
+
 </beans>
