+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/rest/ArtifactRestlet.java
@@ -1,39 +1,24 @@
 package org.carlspring.strongbox.rest;
 
-import org.apache.maven.artifact.Artifact;
-import org.apache.maven.index.ArtifactInfo;
 import org.carlspring.maven.commons.util.ArtifactUtils;
-import org.carlspring.strongbox.io.MultipleDigestInputStream;
-import org.carlspring.strongbox.resource.ResourceCloser;
-import org.carlspring.strongbox.util.ArtifactFileUtils;
-import org.carlspring.strongbox.security.encryption.EncryptionConstants;
 import org.carlspring.strongbox.security.jaas.authentication.AuthenticationException;
-import org.carlspring.strongbox.storage.DataCenter;
-import org.carlspring.strongbox.storage.checksum.ChecksumCacheManager;
-import org.carlspring.strongbox.storage.indexing.RepositoryIndexManager;
-import org.carlspring.strongbox.storage.indexing.RepositoryIndexer;
-import org.carlspring.strongbox.storage.repository.Repository;
 import org.carlspring.strongbox.storage.resolvers.ArtifactResolutionException;
-import org.carlspring.strongbox.storage.resolvers.ArtifactResolutionService;
-import org.carlspring.strongbox.storage.validation.version.VersionValidationException;
-import org.carlspring.strongbox.storage.validation.version.VersionValidator;
-import org.carlspring.strongbox.storage.validation.version.VersionValidatorService;
-import org.carlspring.strongbox.util.MessageDigestUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import org.carlspring.strongbox.storage.resolvers.ArtifactStorageException;
+import org.carlspring.strongbox.storage.services.ArtifactManagementService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.Response;
-import java.io.*;
-import java.security.MessageDigest;
+import java.io.IOException;
+import java.io.InputStream;
 import java.security.NoSuchAlgorithmException;
-import java.util.Arrays;
-import java.util.Set;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 /**
  * @author Martin Todorov
@@ -47,19 +32,7 @@ public class ArtifactRestlet
     private static final Logger logger = LoggerFactory.getLogger(ArtifactRestlet.class);
 
     @Autowired
-    private ChecksumCacheManager checksumCacheManager;
-
-    @Autowired
-    private ArtifactResolutionService artifactResolutionService;
-
-    @Autowired
-    private VersionValidatorService versionValidatorService;
-
-    @Autowired
-    private DataCenter dataCenter;
-
-    @Autowired
-    private RepositoryIndexManager repositoryIndexManager;
+    private ArtifactManagementService artifactManagementService;
 
 
     @PUT
@@ -74,179 +47,19 @@ public class ArtifactRestlet
                    AuthenticationException,
                    NoSuchAlgorithmException
     {
-        String protocol = request.getRequestURL().toString().split(":")[0];
+        handleAuthentication(storage, repository, path, headers, request);
 
-        handleAuthentication(storage, repository, path, headers, protocol);
-
-        performRepositoryAcceptanceValidation(repository, path);
-
-        boolean fileIsChecksum = path.endsWith(".md5") || path.endsWith(".sha1");
-        MultipleDigestInputStream mdis = new MultipleDigestInputStream(is,
-                                                                       new String[]{ EncryptionConstants.ALGORITHM_MD5,
-                                                                                     EncryptionConstants.ALGORITHM_SHA1 });
-
-        // If this is not a checksum file, store the file.
-        // If this is a checksum file, keep the hash in a String.
-        ByteArrayOutputStream baos = null;
-        if (fileIsChecksum)
-        {
-            baos = new ByteArrayOutputStream();
-        }
-
-        OutputStream os = null;
         try
         {
-            os = artifactResolutionService.getOutputStream(repository, path);
+            artifactManagementService.store(storage, repository, path, is);
 
-            int readLength;
-            byte[] bytes = new byte[4096];
-
-            while ((readLength = mdis.read(bytes, 0, bytes.length)) != -1)
-            {
-                if (fileIsChecksum)
-                {
-                    // Buffer the checksum for later validation
-                    baos.write(bytes, 0, readLength);
-                    baos.flush();
-                }
-                else
-                {
-                    os.write(bytes, 0, readLength);
-                    os.flush();
-                    // Write the artifact
-                }
-            }
-
-            System.out.println("# Wrote " + readLength + " bytes.");
+            return Response.ok().build();
         }
-        catch (ArtifactResolutionException e)
+        catch (ArtifactStorageException e)
         {
-            e.printStackTrace();
-            throw new WebApplicationException(e, Response.Status.NOT_FOUND);
+            // TODO: Figure out if this is the correct response type...
+            throw new WebApplicationException(e, Response.Status.FORBIDDEN);
         }
-        finally
-        {
-            ResourceCloser.close(os, logger);
-        }
-
-        final String artifactPath = storage + "/" + repository + "/" + path;
-        if (!fileIsChecksum && os != null)
-        {
-            addChecksumsToCacheManager(mdis, artifactPath);
-
-            if (ArtifactFileUtils.isArtifactFile(path))
-            {
-                final RepositoryIndexer indexer = repositoryIndexManager.getRepositoryIndex(repository);
-                if (indexer != null)
-                {
-                    final Artifact artifact = ArtifactUtils.convertPathToArtifact(path);
-                    final File storageBasedir = new File(dataCenter.getStorage(storage).getBasedir());
-                    final File artifactFile = new File(storageBasedir, artifactPath).getCanonicalFile();
-
-                    indexer.addArtifactToIndex(repository, artifactFile, artifact);
-                }
-            }
-        }
-        else
-        {
-            validateUploadedChecksumAgainstCache(baos, artifactPath);
-        }
-
-        return Response.ok().build();
-    }
-
-    private boolean performRepositoryAcceptanceValidation(String repository,
-                                                          String path)
-            throws WebApplicationException
-    {
-        if (!path.contains("/maven-metadata."))
-        {
-            try
-            {
-                Artifact artifact = ArtifactUtils.convertPathToArtifact(path);
-                Repository r = dataCenter.getRepository(repository);
-
-                final Set<VersionValidator> validators = versionValidatorService.getValidators();
-                for (VersionValidator validator : validators)
-                {
-                    validator.validate(r, artifact);
-                }
-            }
-            catch (VersionValidationException e)
-            {
-                throw new WebApplicationException(e, Response.Status.FORBIDDEN);
-            }
-        }
-
-        return true;
-    }
-
-    private void validateUploadedChecksumAgainstCache(ByteArrayOutputStream baos,
-                                                      String artifactPath)
-    {
-        logger.debug("Received checksum: " + baos.toString());
-
-        String artifactBasePath = artifactPath.substring(0, artifactPath.lastIndexOf('.'));
-        String algorithm = null;
-
-        final String checksumExtension = artifactPath.substring(artifactPath.lastIndexOf('.') + 1,
-                                                                artifactPath.length());
-        if (checksumExtension.equalsIgnoreCase(EncryptionConstants.ALGORITHM_MD5))
-        {
-            algorithm = EncryptionConstants.ALGORITHM_MD5;
-        }
-        else if (checksumExtension.equals("sha1"))
-        {
-            algorithm = EncryptionConstants.ALGORITHM_SHA1;
-        }
-        else
-        {
-            // TODO: Should we be doing something about this case?
-            logger.warn("Unsupported checksum type: " + checksumExtension);
-        }
-
-        final boolean matchesCachedChecksum = matchesChecksum(baos, artifactBasePath, algorithm);
-        if (matchesCachedChecksum)
-        {
-            checksumCacheManager.removeArtifactChecksum(artifactBasePath, algorithm);
-        }
-        else
-        {
-            // TODO: Implement event triggering that handles checksums that don't match the upload file.
-        }
-    }
-
-    private boolean matchesChecksum(ByteArrayOutputStream baos,
-                                    String artifactBasePath,
-                                    String algorithm)
-    {
-        String checksum = baos.toString();
-        String cachedChecksum = checksumCacheManager.getArtifactChecksum(artifactBasePath, algorithm);
-
-        if (cachedChecksum.equals(checksum))
-        {
-            logger.debug("The received " + algorithm + " checksum matches cached one! " + checksum);
-            return true;
-        }
-        else
-        {
-            logger.debug("The received " + algorithm + " does not match cached one! " + checksum + "/" +
-                         cachedChecksum);
-            return false;
-        }
-    }
-
-    private void addChecksumsToCacheManager(MultipleDigestInputStream mdis,
-                                            String artifactPath)
-    {
-        MessageDigest md5Digest = mdis.getMessageDigest(EncryptionConstants.ALGORITHM_MD5);
-        MessageDigest sha1Digest = mdis.getMessageDigest(EncryptionConstants.ALGORITHM_SHA1);
-
-        String md5 = MessageDigestUtils.convertToHexadecimalString(md5Digest);
-        String sha1 = MessageDigestUtils.convertToHexadecimalString(sha1Digest);
-
-        checksumCacheManager.addArtifactChecksum(artifactPath, EncryptionConstants.ALGORITHM_MD5, md5);
-        checksumCacheManager.addArtifactChecksum(artifactPath, EncryptionConstants.ALGORITHM_SHA1, sha1);
     }
 
     @GET
@@ -264,8 +77,7 @@ public class ArtifactRestlet
     {
         logger.debug(" repository = " + repository + ", path = " + path);
 
-        String protocol = request.getRequestURL().toString().split(":")[0];
-        handleAuthentication(storage, repository, path, headers, protocol);
+        handleAuthentication(storage, repository, path, headers, request);
 
         if (!ArtifactUtils.isArtifact(path))
         {
@@ -275,7 +87,7 @@ public class ArtifactRestlet
         InputStream is;
         try
         {
-            is = artifactResolutionService.getInputStream(repository, path);
+            is = artifactManagementService.resolve(storage, repository, path);
         }
         catch (ArtifactResolutionException e)
         {
@@ -297,35 +109,14 @@ public class ArtifactRestlet
 
         try
         {
-            artifactResolutionService.delete(repository, path);
-
-            final RepositoryIndexer indexer = repositoryIndexManager.getRepositoryIndex(repository);
-            if (indexer != null)
-            {
-                final Artifact a = ArtifactUtils.convertPathToArtifact(path);
-                indexer.delete(Arrays.asList(new ArtifactInfo(repository,
-                                                              a.getGroupId(),
-                                                              a.getArtifactId(),
-                                                              a.getVersion(),
-                                                              a.getClassifier())));
-            }
+            artifactManagementService.delete(repository, path);
         }
-        catch (ArtifactResolutionException e)
+        catch (ArtifactStorageException e)
         {
             throw new WebApplicationException(e, Response.Status.NOT_FOUND);
         }
 
         return Response.ok().build();
-    }
-
-    public ChecksumCacheManager getChecksumCacheManager()
-    {
-        return checksumCacheManager;
-    }
-
-    public void setChecksumCacheManager(ChecksumCacheManager checksumCacheManager)
-    {
-        this.checksumCacheManager = checksumCacheManager;
     }
 
 }
